requirements.txt
# Core
langchain==0.3.25
langgraph==0.2.74
openai==1.82.0

# Data & Validation
pydantic==2.11.4
pandas==2.2.3
httpx==0.28.1

# Database
supabase==2.15.2

# API
fastapi==0.115.12
uvicorn==0.34.2

# Utilities
python-dotenv==1.1.0
loguru==0.7.3
tenacity==9.1.2